requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
apscheduler>=3.10.0
requests>=2.31.0
pytest>=7.4.0
